+++ src/main/java/com/trackery/trackeryfrontserver/config/SecurityConfig.java
@@ -33,7 +33,7 @@ public class SecurityConfig {
 		http
 			.csrf(csrf -> csrf.disable())
 			.authorizeHttpRequests(auth -> auth
-				.requestMatchers("/", "/css/**", "/js/**", "/images/**", "/error", "/proxy/**").permitAll()
+				.requestMatchers("/", "/css/**", "/js/**", "/images/**", "/error", "/api/**").permitAll()
 				.anyRequest().authenticated()
 			)
 			.formLogin(login -> login.disable())
