--- src/main/resources/application.properties
@@ -1,7 +1,0 @@
-spring.application.name=Trackery-FrontServer
-logging.level.org.springframework.security=DEBUG
-
-server.port=${SERVER_PORT}
-spring.thymeleaf.cache=false
-
-api.server.url=${BASE_PATH}
