--- src/test/java/com/trackery/trackeryfrontserver/TrackeryFrontServerApplicationTests.java
@@ -1,13 +1,0 @@
-package com.trackery.trackeryfrontserver;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TrackeryFrontServerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
